% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoboot.R
\name{blockboot}
\alias{blockboot}
\title{Compute Optimal Block Length for Non-overlapping, Overlapping, synthetic-overlapping, appended-overlapping and Circular Block Bootstrap}
\usage{
blockboot(
  ts,
  R,
  methods = c("optnbb", "optmbb", "optcbb", "opttmbb", "opttcbb")
)
}
\arguments{
\item{ts}{univariate time series}

\item{R}{number of resample}

\item{methods}{"optnbb", "optmbb", "optcbb", "opttmbb", "opttcbb"}
}
\description{
More detailed Description
}
\section{Functions}{
\itemize{
\item \code{blockboot}: package helps to obtain the optimal block length of a time series data
}}

