% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoboot.R
\name{blockboot}
\alias{blockboot}
\title{OBL: Optimal Block Length
 @description Compute Optimal Block Length for Non-overlapping, Overlapping, Circular Block, tapered moving, and tapered circular Block Bootstrap method

\code{blockboot} returns the a data frame with five(5) rows and three(3) columns.}
\usage{
blockboot(
  ts,
  R,
  seed,
  n_cores,
  methods = c("optnbb", "optmbb", "optcbb", "opttmbb", "opttcbb")
)
}
\arguments{
\item{ts}{univariate time series}

\item{R}{number of resample}

\item{seed}{RNG seed}

\item{n_cores}{number of core(s) to be used on your operaterating system}

\item{methods}{"optnbb", "optmbb", "optcbb", "opttmbb", "opttcbb"}
}
\description{
OBL: Optimal Block Length
 @description Compute Optimal Block Length for Non-overlapping, Overlapping, Circular Block, tapered moving, and tapered circular Block Bootstrap method

\code{blockboot} returns the a data frame with five(5) rows and three(3) columns.
}
\section{Functions}{
\itemize{
\item \code{blockboot}: package helps to obtain the optimal block length of a time series data
}}

